test(openshell): preserve remote seed fixture ownership
